requirements.txt
asgiref==3.5.2
autopep8==2.0.0
boto3==1.26.29
botocore==1.29.29
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-extensions==3.2.1
django-storages==1.13.1
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
MarkupSafe==2.1.1
numpy==1.24.1
pandas==1.5.2
Pillow==9.3.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==22.1.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
